requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch>=2.9.0
torchvision>=0.24.0
transformers @ git+https://github.com/huggingface/transformers@3a8d291ac439fca3decb9b800ea22500cffeaf50
numpy>=2.0.0
opencv-python>=4.10.0
Pillow>=10.0.0
matplotlib>=3.9.0
PyYAML>=6.0
scipy>=1.13.0
